Refactor tests

+++ api/admin/folders_test.go
@@ -1,14 +1,16 @@
-package admin
+package admin_test
 
 import (
 	"testing"
 	"time"
+
+	"github.com/cloudinary/cloudinary-go/api/admin"
 )
 
 const testFolder = "000-go-folder"
 
 func TestFolders_CreateFolder(t *testing.T) {
-	resp, err := adminApi.CreateFolder(ctx, CreateFolderParams{Folder: testFolder})
+	resp, err := adminApi.CreateFolder(ctx, admin.CreateFolderParams{Folder: testFolder})
 
 	if err != nil || resp.Success != true {
 		t.Error(resp, err)
@@ -16,7 +18,7 @@ func TestFolders_CreateFolder(t *testing.T) {
 }
 
 func TestFolders_DeleteFolder(t *testing.T) {
-	resp, err := adminApi.DeleteFolder(ctx, DeleteFolderParams{Folder: testFolder})
+	resp, err := adminApi.DeleteFolder(ctx, admin.DeleteFolderParams{Folder: testFolder})
 
 	if err != nil || len(resp.Deleted) < 1 {
 		t.Error(resp, err)
@@ -24,7 +26,7 @@ func TestFolders_DeleteFolder(t *testing.T) {
 }
 
 func TestFolders_RootFolders(t *testing.T) {
-	resp, err := adminApi.RootFolders(ctx, RootFoldersParams{MaxResults: 5})
+	resp, err := adminApi.RootFolders(ctx, admin.RootFoldersParams{MaxResults: 5})
 
 	if err != nil || resp.TotalCount < 1 {
 		t.Error(resp, err)
@@ -32,24 +34,24 @@ func TestFolders_RootFolders(t *testing.T) {
 }
 
 func TestFolders_SubFolders(t *testing.T) {
-	cfResp, err := adminApi.CreateFolder(ctx, CreateFolderParams{Folder: testFolder})
+	cfResp, err := adminApi.CreateFolder(ctx, admin.CreateFolderParams{Folder: testFolder})
 	if err != nil || cfResp.Success != true {
 		t.Error(cfResp, err)
 	}
 
-	cfResp, err = adminApi.CreateFolder(ctx, CreateFolderParams{Folder: testFolder + "/" + testFolder})
+	cfResp, err = adminApi.CreateFolder(ctx, admin.CreateFolderParams{Folder: testFolder + "/" + testFolder})
 	if err != nil || cfResp.Success != true {
 		t.Error(cfResp, err)
 	}
 
 	time.Sleep(1 * time.Second)
 
-	resp, err := adminApi.RootFolders(ctx, RootFoldersParams{MaxResults: 1})
+	resp, err := adminApi.RootFolders(ctx, admin.RootFoldersParams{MaxResults: 1})
 	if err != nil || resp == nil || resp.TotalCount < 1 {
 		t.Error(resp, err)
 	}
 
-	resp, err = adminApi.SubFolders(ctx, SubFoldersParams{Folder: resp.Folders[0].Path, MaxResults: 2})
+	resp, err = adminApi.SubFolders(ctx, admin.SubFoldersParams{Folder: resp.Folders[0].Path, MaxResults: 2})
 	if err != nil || resp.TotalCount < 1 {
 		t.Error(resp, err)
 	}
