+++ api/admin/streaming_profiles_test.go
@@ -1,7 +1,9 @@
-package admin
+package admin_test
 
 import (
 	"testing"
+
+	"github.com/cloudinary/cloudinary-go/api/admin"
 )
 
 const SPName = "00-go-sp"
@@ -21,7 +23,7 @@ func TestStreamingProfiles_Get(t *testing.T) {
 		t.Error(lResp, err)
 	}
 
-	resp, err := adminApi.GetStreamingProfile(ctx, GetStreamingProfileParams{Name: lResp.Data[0].Name})
+	resp, err := adminApi.GetStreamingProfile(ctx, admin.GetStreamingProfileParams{Name: lResp.Data[0].Name})
 
 	if err != nil {
 		t.Error(resp, err)
@@ -29,10 +31,10 @@ func TestStreamingProfiles_Get(t *testing.T) {
 }
 
 func TestStreamingProfiles_Create(t *testing.T) {
-	resp, err := adminApi.CreateStreamingProfile(ctx, CreateStreamingProfileParams{
+	resp, err := adminApi.CreateStreamingProfile(ctx, admin.CreateStreamingProfileParams{
 		Name:            SPName,
 		DisplayName:     "Go SP",
-		Representations: StreamingProfileRepresentations{{Transformation: "c_fill,w_1000,h_1000"}},
+		Representations: admin.StreamingProfileRepresentations{{Transformation: "c_fill,w_1000,h_1000"}},
 	})
 
 	if err != nil || resp.Error.Message != "" {
@@ -41,10 +43,10 @@ func TestStreamingProfiles_Create(t *testing.T) {
 }
 
 func TestStreamingProfiles_Update(t *testing.T) {
-	resp, err := adminApi.UpdateStreamingProfile(ctx, UpdateStreamingProfileParams{
+	resp, err := adminApi.UpdateStreamingProfile(ctx, admin.UpdateStreamingProfileParams{
 		Name:            SPName,
 		DisplayName:     "Go SP Updated",
-		Representations: []RawStreamingProfileRepresentation{{"c_fill,w_1001,h_1001"}},
+		Representations: admin.StreamingProfileRepresentations{{"c_fill,w_1001,h_1001"}},
 	})
 
 	if err != nil || resp.Data.DisplayName != "Go SP Updated" {
@@ -53,7 +55,7 @@ func TestStreamingProfiles_Update(t *testing.T) {
 }
 
 func TestStreamingProfiles_Delete(t *testing.T) {
-	resp, err := adminApi.DeleteStreamingProfile(ctx, DeleteStreamingProfileParams{Name: SPName})
+	resp, err := adminApi.DeleteStreamingProfile(ctx, admin.DeleteStreamingProfileParams{Name: SPName})
 
 	if err != nil || resp.Message != "deleted" {
 		t.Error(resp, err)
