+++ api/admin/upload_mappings_test.go
@@ -1,7 +1,9 @@
-package admin
+package admin_test
 
 import (
 	"testing"
+
+	"github.com/cloudinary/cloudinary-go/api/admin"
 )
 
 const UMFolder = "wiki"
@@ -9,7 +11,7 @@ const UMTemplate = "https://u.wiki.example.com/wiki-images/"
 const UMTemplateUpdated = "https://images.example.com/product_assets/images/"
 
 func TestUploadMappings_Create(t *testing.T) {
-	resp, err := adminApi.CreateUploadMapping(ctx, CreateUploadMappingParams{
+	resp, err := adminApi.CreateUploadMapping(ctx, admin.CreateUploadMappingParams{
 		Folder:   UMFolder,
 		Template: UMTemplate,
 	})
@@ -20,7 +22,7 @@ func TestUploadMappings_Create(t *testing.T) {
 }
 
 func TestUploadMappings_List(t *testing.T) {
-	resp, err := adminApi.ListUploadMappings(ctx, ListUploadMappingsParams{})
+	resp, err := adminApi.ListUploadMappings(ctx, admin.ListUploadMappingsParams{})
 
 	if err != nil || len(resp.Mappings) < 1 {
 		t.Error(resp, err)
@@ -28,13 +30,13 @@ func TestUploadMappings_List(t *testing.T) {
 }
 
 func TestUploadMappings_Get(t *testing.T) {
-	lResp, err := adminApi.ListUploadMappings(ctx, ListUploadMappingsParams{})
+	lResp, err := adminApi.ListUploadMappings(ctx, admin.ListUploadMappingsParams{})
 
 	if err != nil || lResp.Error.Message != "" {
 		t.Error(lResp, err)
 	}
 
-	resp, err := adminApi.GetUploadMapping(ctx, GetUploadMappingParams{Folder: lResp.Mappings[0].Folder})
+	resp, err := adminApi.GetUploadMapping(ctx, admin.GetUploadMappingParams{Folder: lResp.Mappings[0].Folder})
 
 	if err != nil {
 		t.Error(resp, err)
@@ -42,7 +44,7 @@ func TestUploadMappings_Get(t *testing.T) {
 }
 
 func TestUploadMappings_Update(t *testing.T) {
-	resp, err := adminApi.UpdateUploadMapping(ctx, UpdateUploadMappingParams{
+	resp, err := adminApi.UpdateUploadMapping(ctx, admin.UpdateUploadMappingParams{
 		Folder:   UMFolder,
 		Template: UMTemplateUpdated,
 	})
@@ -53,7 +55,7 @@ func TestUploadMappings_Update(t *testing.T) {
 }
 
 func TestUploadMappings_Delete(t *testing.T) {
-	resp, err := adminApi.DeleteUploadMapping(ctx, DeleteUploadMappingParams{Folder: UMFolder})
+	resp, err := adminApi.DeleteUploadMapping(ctx, admin.DeleteUploadMappingParams{Folder: UMFolder})
 
 	if err != nil || resp.Message != "deleted" {
 		t.Error(resp, err)
